+++ tests/ZfcTwitterBootstrap/View/Helper/NavigationTest.php
@@ -91,50 +91,4 @@ class NavigationTest extends \PHPUnit_Framework_TestCase
         $actual = $this->helper->render($container);
         $this->assertEquals($expected, $actual);
     }
-
-    public function testBreadcrumbs()
-    {
-        $this->helper->setDefaultProxy('ztbbreadcrumbs');
-
-        $container = new \Zend\Navigation\Navigation(array(
-            array(
-                'label' => 'Page 1',
-                'id'    => 'p1',
-                'uri'   => 'p1',
-            ),
-            array(
-                'label' => 'Page 2',
-                'id'    => 'p2',
-                'uri'   => 'p2',
-                'pages' => array(
-                    array(
-                        'label' => 'Page 2.1',
-                        'id'    => 'p2-1',
-                        'uri'   => 'p2-1',
-                    ),
-                    array(
-                        'label' => 'Page 2.2',
-                        'id'    => 'p2-2',
-                        'uri'   => 'p2-2',
-                        'pages' => array(
-                            array(
-                                'label' => 'Page 2.2.1',
-                                'id'    => 'p2-2-1',
-                                'uri'   => 'p2-2-1',
-                                'active' => true,
-                            ),
-                        ),
-                    ),
-                ),
-            ),
-        ));
-
-        $expected = '<ul class="breadcrumb">'
-                  . '<li>Page 2 <span class="divider">/</span></li>'
-                  . '<li>Page 2.2 <span class="divider">/</span></li>'
-                  . '<li class="active">Page 2.2.1</li></ul>';
-
-        $actual = $this->helper->render($container);
-        $this->assertEquals($expected, $actual);
-    }
 }
